Use mockall to test disclosure events in unit test
The `Wallet` disclosure unit tests checked for the correct disclosure
events my manually making these events, then returning them from
`MockStorage`. This made it so that the test code was not testing the
actual implementation, but the test code itself.

This changes all `Wallet` disclosure unit tests so that the correct
event expectations are set on the `MockStorage` using mockall.