PVW-5091: Handle PIN confirmation failures in transfer BLoC
* Add explicit `BuildContext` type

Explicitly add `BuildContext` type to the `create` function's context parameter in `BlocProvider` and use this instance of the BuildContext to fetch the dependencies.

* Pull up PinBloc (above WalletTransferSource) so UI can react

While the wallet_core is busy verifying the pin (and transferring the wallet db) the user could press
the back button, causing the UI to enter an incorrect state. The visibility of the back button / ability
to pop is now conditional, taking the pin validation state into account.

* Handle PIN confirmation failures in transfer BLoC

The handler for WalletTransferPinConfirmationFailed was not set in the wallet_transfer_source_bloc, this
is a programming mistake and fixed with this commit. Furthermore a generic error, instead of a more accurate
WalletTransferNetworkError was exposed, the code has been update to now emit this more specific network error.