PVW-5094: Handle exceptions during wallet transfer status polling
* Handle exceptions during wallet transfer status polling

Fetching the TransferStatus can return an error when trying to do so when no transfer_session is active, this could happen when
the user decided to skip the transfer on the destination device, after initially opting in. (the initial opt in would start the polling).

We now handle this gracefully by catching the error and returning (thereby ending the stream).