PVW-5114: Show 'transferring' state on destination wallet
* Fix conflict resolve issue & doc

- A code bloc was incorrectly left out during the conflict resolvement, this is now re-introduced
- Fix a minor typo in transfer state dock

* Fix test that verifies back press is ignored

* Reduce wallet transfer status polling delay

The polling delay for the wallet transfer status has been reduced from 3 seconds to 2 seconds.

* Use `ReceiveWalletTransferUseCase` to download wallet

This commit refactors the wallet transfer process on the target device so that it moves into the transferring state and starts downloading the wallet_transfer payload once the status
informs us that the payload is ready to be downloaded. The polling of the wallet_transfer status stops as soon as the download begins.

* Implement `receiveWalletTransfer` from core

This change introduces the `receiveWalletTransfer` functionality, connecting it from the wallet core up to the application's dependency injection and use case layers.

It includes:
- Adding `receiveWalletTransfer` to the `TypedWalletCore` and repository abstractions.
- Implementing a new `ReceiveWalletTransferUseCase` to expose the functionality to the domain layer.
- Providing a mock implementation for `receiveWalletTransfer`.

* Offer separate receive_wallet_payload endpoint in the Flutter API in order to show the waiting for move screen