PVW-5055: Navigate to splash screen onPop (transfer)
* Navigate to splash screen onPop

On Android the user could close the app by pressing the back button from the 'transfer success' page on the source.
This was inconsistent with the iOS behaviour as that only allowed to restart the app, android now behaves in the same
way: that is navigating to the splash screen (i.e. app restart) on a back gesture or button press.