Start transfer on pinConfirmed event and keep polling state
The `startWalletTransferUseCase` is now invoked immediately after the user confirms their PIN, rather than waiting for the `readyForTransferConfirmed` status from the transfer status observer. Additionally we no longer cancel the stream that polls the wallet provider for the current state, as the stream will notify
us once the target device has completed the transfer so we can move to the success state on the source.