Refactor: Rename WalletTransferStatus to TransferSessionState
This commit renames the `WalletTransferStatus` enum to `TransferSessionState` to align with the naming in the wallet core. Also the exposed states are no longer directly tied to the corresponding ui state, but instead
copy the naming convention by the core, where the docs of this enum can be explored to see what UI state should
be associated to it.