PVW-5135: [Flutter] Handle Denied- & WrongDigid errors
* Fix lint warning

* Improve CoreErrorMapper Test

Add tests to `CoreErrorMapper` to verify the mapping of additional `FlutterApiError` types to their corresponding `CoreError` types.

* Verify the behaviour of CoreErrorExtension

* Add tests for ApplicationError subclasses

Adds unit tests for the `ApplicationError` subclasses. The tests verify that the `props` for `Equatable` are correctly implemented for each error type.

This also adds `Equatable` to `CheckPinResult` to make testing `CheckPinError` possible.

* Verify handling of DigiD errors in PID issuance/renewal and PIN recovery flows

Add tests to verify that the correct states are emitted when `WrongDigidError` or `DeniedDigidError` occur during the following processes:
- PID renewal
- Initial PID issuance (personalization)
- PIN recovery

* Handle WrongDigidError and DeniedDigidError in renew PID flow

* Handle cancelled DigiD login during PIN recovery

* Handle DigiD denial during personalization

Handle the `DeniedDigidError` error in the wallet personalization flow by emitting WalletPersonalizeLoginWithDigidFailed(cancelled = true) when this error is detected.

* Parse and expose DeniedDigidError

The wallet core can now throw a DeniedDigiDError, this commit makes it so that the wallet_app can
recognize it and thus any downstream components can start reacting to it.