PVW-5091: Fix state related transfer issues
* Migrate to Stream.periodic for transfer session state polling

Instead of the yielding method that internally awaited a Future.delayed to implement the 2-second polling mechanism, we now
rely on a Stream.periodic stream in which we asyncMap to get the latest transfer session state. This fixes an issue where the
yielding async* method did not realize the observer was gone, causing the app to crash when the walletTransferState() threw
an exception.

Furthermore the mapping from an error into an ApplicationError has been consolidated into the static exceptionToApplicationError().

* Cancel wallet transfer when returning to dashboard (source)

When the user is transferring a wallet (acting as the source device) and navigates back to the dashboard, the wallet transfer process is now automatically cancelled.