PVW-5129: Always render child in PinOverlay to maintain state
* Make SetupSecurityScreen a secured route

This makes sure the 'PinOverlay' is shown when the wallet is locked after setting up the pin (but not yet navigating to the next step of the setup process).
This works (and doesn't block the setup process) because the pin overlay also checks if the wallet is initialized before rendering the overlay, and since the
wallet is not initialized before setting up the pin the normal flow is not blocked by the PinOverlay.

* Refactor: Remove LockStateMixin

The `LockStateMixin` and its corresponding tests have been deleted as it is no longer in use.

* Refactor SetupHelper to OnboardingHelper

- The `SetupHelper` utility class has been renamed to `OnboardingHelper` to more accurately reflect its purpose of managing the onboarding flow's steps.
- It's init is now called in the main() app start method, to make sure it's initialized before starting
the app.
- It no longer initializes based on a context, this allows us to initialize it before the app start
- Tests have been updated to verify the new initialization behaviour
- The init() call has been removed from the splashscreen, along with the 'check initialization' call in the splashbloc. Since initialization is now guaranteed at app start.

* Remove (unused) provideMocks flag from DI

This flag was no longer used, since all mock calls are routes through the mocked WalletCore
which is set at app start in main.dart.

* Await SetupHelper initialization on splash screen

Introduced a `Completer` to the `SetupHelper` to ensure its initialization, which determines the number of setup steps, is complete before the splash screen logic proceeds.

This prevents a race condition that occurred when the app was terminated after PIN setup but before PID registration, ensuring the correct navigation path is chosen upon restart.

* Update PinOverlayTests

Verify that:
- Child is rendered below PinOverlay so state can be kept
- PinOverlay is displayed when app is locked
- Child (content) is displayed when app is unlocked

* Remove LockStateMixin from UI components

The `LockStateMixin` and its associated `onLock` logic, which navigated to the dashboard when the app was locked, has been removed from several UI screens:
- MenuScreen
- SettingsScreen
- NeedHelpScreen
- ContactScreen
- ConfirmWithPinScreen
- WalletTransferTargetScreen

These screens have been converted from `StatefulWidget` to `StatelessWidget` where applicable as the mixin was the only reason for them to have state.

* Wrap PinScreen in a PrimaryScrollController

Because the PinScreen can now be rendered on top of any content, we make sure to provide it with a dedicated
ScrollController to avoid conflicts of multiple WalletAppBars trying to attach to the same (default) scrollController.

* Always render child in PinOverlay to maintain state

The child widget is now always rendered in the widget tree, even when the screen is locked. The locked state is rendered in a Stack on top of the child, instead of conditionally replacing it. This prevents the child widget from being unmounted and re-mounted when the application is locked and unlocked, making it so that the bloc (and thus state) is maintained.