PVW-5195: Add tests for Autolocking
* Fix lint issue

* Mock VersionString in ErrorDetailsSheet test

Make sure the goldens have a version string to render.

* Update wallet transfer goldens

* Verify that wallet transfer is canceled (source) when dashboard is shown

* Update golden files

Updates golden files for error details sheet and wallet transfer.

* Centralize Mockito dummy providers in `flutter_test_config.dart`

Removes local `provideDummy` calls from individual test files and consolidates them into a central `_setupMockitoDummies` function in `flutter_test_config.dart`.

This change centralizes test setup, reduces boilerplate code in tests, and ensures consistent dummy data is available for all test scenarios.

* Add CoreError tests for wallet core methods

Adds tests to verify that `CoreError` exceptions are correctly thrown when the underlying FFI calls fail for various methods in `TypedWalletCore`.

* Add tests for `AutoLockService`

* Add tests for AutoLockObserver

Added tests for the `AutoLockObserver` to verify:
- Interaction resets the idle timeout.
- The wallet locks after the configured idle timeout.
- The wallet locks when the app is backgrounded for longer than the timeout.
- The wallet does not lock when the app is backgrounded for less than the timeout.

Also migrated to clock.stopwatch() so that we can control the time in tests.

* Add tests for InteractionDetector

Adds a widget test for the `InteractionDetector` to verify that the `onInteraction` callback is triggered when the widget is tapped.