Update `WalletState` enum with recursive Locked state
The `Locked` state now contains a `substate` field, indicating the state the wallet was in before it was locked (and thus also the
state the wallet will be in after it is unlocked). It also introduces an 'Emtpy' state, indicating the user has registered a pin, but not yet added the pid.