Optimize navigate to transfer screen from splash
Instead of navigating from the dashboard after unlock, the app now navigates directly to the wallet transfer screen from the splash screen when the wallet state is `TransferPossible` or `Transferring` as a target.
This resolves a UI glitch where the user could briefly see the dashboard before the transfer was re-launched. The 'transfer' stopped dialog is now managed by the WalletTransferEventListener to make sure it does
not show up before the user unlocks the app.

This change also reduces the complexity in the wallet_transfer_target_screen.dart so it can be made Stateless. It also reduces the complexity of the wallet_transfer_event_listener.dart since it no longer has to
handle the navigation requests, it only needs to notify the user about the transfer being stopped.