Add lock screen handling to PIN recovery
The PIN recovery screen is converted to a `StatefulWidget` and uses the `LockStateMixin` to handle app locking and unlocking.
When the app is locked during PIN recovery, the user is redirected to the dashboard, which implicitly means that the pin recovery
progress is reset and the user is back on the 'unlock with pin' screen. From here the user can recovery the pin again, or simply
enter the pin to unlock the app and navigate to the dashboard.