PVW-5271: Fix 'recover pin' related bugs
* Fix CancelPinRecovery issues

- Also check the locked state, which can occur during the pin recovery flow when the user
navigates back to the app (i.e. does not complete the digid web login).
- Pop back to dashboard from the stopped screen, as it can be that there is no underlying
route when coming from the splash.

* Change error description for pin recovery session errors

* Fix: Correct cancel invocation count in issuance test

* Auto cancel session on RenewPidBloc closure

Also fix incorrect issuance flow check inside the core_pid_repository.dart

* Fix 'recover pin' related bugs

- Cancel disclosure / issuance sessions in onClose callbacks, this makes sure
the wallet is no longer in an issuance/disclosure state when the recover pin
screen is opened. This relies on navigating to the recover pin screen as a
replacement (i.e. popping the disclosure/issuance screen so the corresponding
bloc is closed.
- Update CancelPinRecoveryUseCase to check the state before cancelling, avoiding
a state error when trying to cancel an non-existing recovery session.
- Remove stale queued request (i.e. we now drop queued request if a newer deeplink event comes in)