PVW-3446: RDO MAX unmodified upstream enablement
* More sanity checks when using existing repo
* Make sure that an already checked out rdo-max is at the requested branch tag
* Removed redis helm-chart, k8s deployment comes in follow-up MR
* Further cleanup of build.yml
* Specify a couple of build arguments
* Run on our linux host, restore standard credential configuration
* Alternative way to pass credentials
* Add BUILDAH_ISOLATION=chroot
* Don't specify quay.io when we go through quay-proxy
* Use buildah unprivileged/rootless to build image in CI
* Set RDO_MAX_VERSION as image tag
* Added sha256sum + checking, artifact creation
* Use specific version of nl-rdo-max, use curl to fetch, artifact paths added
* Added path to artifacts
* Added rules
* Need to invoke npm command before build, split into two jobs
* Added digid-connector trigger and test-sub-build yaml
* Script uses bashisms, should use /bin/busybox sh
* Using app as docker-compose service name
* Use new make targets, container name
* Removed commit specification for rdo-max, Renamed _BASE_TAG to _VERSION, which can be both a version tag or a branch name
* Removed local patches for MAX
* Use upstream unmodified MAX

We currently use a specific pinned version of MAX, which we then patch
to remove the ui-theme, add docker and compose files, add a create-jwk
target and fix a primary idp showing on the redirect page.

Some of these things (like showing the primary idp on the redirect page)
have been fixed in later (3.x+) version of MAX. Others we are working
to upstream.

This commit adjusts `setup-devenv.sh` as follows:

  * Don't git-clone a specific commit anymore, get the latest public max
  * Don't apply any local patches anymore, run unmodified upstream
  * Explicitly uninstall the nl-rijksoverheid-ui-theme
  * Pass current UID/GID to docker compose build and buildah commands
  * Use new create-jwk target which uses a new upstream script

Don't unconditionally rebuild MAX (digid-connector) on start

We used to rebuild MAX in `startup-devenv.sh`. This is not necessary as
the container is built in the `setup-devenv.sh` stage. We now simply
compose up with detach instead of build and force-recreate.

Add client_authentication_method setting for max

In `client.json`, MAX 3.x requires a new setting for specifying the
client authentication method called `client_authentication_method`.
This setting does not have a default setting, so MAX errors out when
you don't set it. We set it to `none`.